third_party/blink/web_tests/svg/dynamic-updates/SVGTextElement-dom-dy-attr-expected.txt

SVG 1.1 dynamic update tests
Text content

Tests dynamic updates of the 'dy' attribute of the SVGTextElement object

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS textElement.getAttribute('dy') is "0"
PASS textElement.getAttribute('dy') is "-30"
PASS successfullyParsed is true

TEST COMPLETE
